pyproject.toml
[project]
name = "mail"
authors = [
    { name = "Frappe Technologies Pvt. Ltd.", email = "developers@frappe.io"}
]
description = "Mail"
requires-python = ">=3.10"
readme = "README.md"
dynamic = ["version"]
dependencies = [
    "dnspython~=2.4.2",
    "dkimpy~=1.1.5",
    "pyspf~=2.0.14",
]

[build-system]
requires = ["flit_core >=3.4,<4"]
build-backend = "flit_core.buildapi"

# These dependencies are only installed when developer mode is enabled
[tool.bench.dev-dependencies]
# package_name = "~=1.1.0"
